chore: move data-generation for dev usage scripts into specific directory
